Inheritance - Creating new classes
"Inheritance in Object-Oriented Programming can be described as a process of creating new classes from existing classes. ... An existing class that is "parent" of a new class is called a base class. New class that inherits properties of the base class is called a derived class. Inheritance is a technique of code reuse."
                  - https://www.tutorialcup.com/cplusplus