Storage of data elements
Array, vector, list, forward-list, deque.  Storing data sequentially but accessing it FIFO or LIFO, with push or pop, and with the deque, placing data at either the  front or the back of the same list.  Also, using the deque, accessing specific data elements regardless of their location in the list.